finance-prediction-system/backend/data_collector/requirements.txt
psycopg2-binary==2.9.5
redis==4.5.1
requests==2.28.2
python-dotenv==1.0.0
APScheduler==3.10.1
pytz==2023.3
